+++ ios/StripeSdkImpl.swift
@@ -384,11 +384,16 @@ public class StripeSdkImpl: NSObject, UIAdaptivePresentationControllerDelegate {
             if paymentMethodType == .USBankAccount && paymentMethodData == nil {
                 return STPSetupIntentConfirmParams(clientSecret: setupIntentClientSecret, paymentMethodType: .USBankAccount)
             } else {
-                let factory = PaymentMethodFactory.init(paymentMethodData: paymentMethodData, options: options, cardFieldView: cardFieldView, cardFormView: cardFormView)
+                let factory = PaymentMethodFactory.init(paymentMethodData: paymentMethodData, options: options, cardFieldView: cardFieldView, cardFormView: cardFormView, publishableKey: STPAPIClient.shared.publishableKey ?? "")
                 let parameters = STPSetupIntentConfirmParams(clientSecret: setupIntentClientSecret)
 
                 if let paymentMethodId = paymentMethodData?["paymentMethodId"] as? String {
                     parameters.paymentMethodID = paymentMethodId
+                    do {
+                        parameters.paymentMethodOptions = try factory.createOptions(paymentMethodType: paymentMethodType)
+                    } catch {
+                        err = Errors.createError(ErrorType.Failed, error as NSError?)
+                    }
                 } else {
                     do {
                         parameters.paymentMethodParams = try factory.createParams(paymentMethodType: paymentMethodType)
@@ -621,7 +626,8 @@ public class StripeSdkImpl: NSObject, UIAdaptivePresentationControllerDelegate {
             paymentMethodData: params["paymentMethodData"] as? NSDictionary,
             options: options,
             cardFieldView: cardFieldView,
-            cardFormView: cardFormView
+            cardFormView: cardFormView,
+            publishableKey: STPAPIClient.shared.publishableKey ?? ""
         )
         do {
             paymentMethodParams = try factory.createParams(paymentMethodType: paymentMethodType)
@@ -964,7 +970,7 @@ public class StripeSdkImpl: NSObject, UIAdaptivePresentationControllerDelegate {
                 return STPPaymentIntentParams(clientSecret: paymentIntentClientSecret, paymentMethodType: .USBankAccount)
             } else {
                 guard let paymentMethodType = paymentMethodType else { return STPPaymentIntentParams(clientSecret: paymentIntentClientSecret) }
-                let factory = PaymentMethodFactory.init(paymentMethodData: paymentMethodData, options: options, cardFieldView: cardFieldView, cardFormView: cardFormView)
+                let factory = PaymentMethodFactory.init(paymentMethodData: paymentMethodData, options: options, cardFieldView: cardFieldView, cardFormView: cardFormView, publishableKey: STPAPIClient.shared.publishableKey ?? "")
                 let paymentMethodId = paymentMethodData?["paymentMethodId"] as? String
                 let parameters = STPPaymentIntentParams(clientSecret: paymentIntentClientSecret)